requirements.txt
fastapi==0.73.0
numpy==1.19.5
uvicorn==0.15.0
image==1.5.33
tensorflow-cpu==2.7.0
protobuf==3.20.1
firebase-admin==5.2.0
pydantic==1.9.0
